fix(gpu): AQC-734: promote critical GPU accounting fields from f32 to f64
Change GpuComboState accumulator fields (balance, total_pnl, total_fees,
gross_profit, gross_loss, max_drawdown, peak_equity) from float to double.
This prevents precision loss over 10K+ trade accumulations.

- CUDA: GpuComboState uses double for 7 accumulator fields
- CUDA: apply_close/apply_partial_close compute PnL/fee as double
- CUDA: equity tracking uses double for drawdown computation
- CUDA: result packing casts double→float for GpuResult output
- CUDA: entry sizing explicitly casts (float)state.balance
- Rust: buffers.rs mirrors f64 layout, removes _state_pad, adds _acc_pad
- Rust: gpu_host.rs/tpe_sweep.rs cast initial_balance to f64
- Struct size: 4000→4032 bytes (16-byte aligned, verified by compile-time assert)

+++ backtester/crates/bt-gpu/src/gpu_host.rs
@@ -262,8 +262,8 @@ impl BatchBuffers {
 
         let mut states_host = vec![GpuComboState::zeroed(); configs.len()];
         for s in &mut states_host {
-            s.balance = initial_balance;
-            s.peak_equity = initial_balance;
+            s.balance = initial_balance as f64;
+            s.peak_equity = initial_balance as f64;
         }
         let states = ds.dev.htod_sync_copy(&states_host).unwrap();
         let results = ds.dev.alloc_zeros::<GpuResult>(configs.len()).unwrap();
@@ -323,8 +323,8 @@ impl BatchBuffers {
 
         let mut states_host = vec![GpuComboState::zeroed(); configs.len()];
         for s in &mut states_host {
-            s.balance = initial_balance;
-            s.peak_equity = initial_balance;
+            s.balance = initial_balance as f64;
+            s.peak_equity = initial_balance as f64;
         }
         let states = ds.dev.htod_sync_copy(&states_host).unwrap();
         let results = ds.dev.alloc_zeros::<GpuResult>(configs.len()).unwrap();
